pyproject.toml
[project]
name = "my-agent"
version = "0.1.0"
description = "Intelligent API agent powered by Google Gemini 2.0 Flash with conversation context and structured responses"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.39.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-google-genai>=2.1.12",
    "pydantic>=2.11.9",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = [
    "*.py",
    "tools/*.py",
]
